refactor: CRUD completo y probado en user y category. - Se agrego paquete DTO para mejor manejo de entrada y respuesta de las peticiones http.

+++ src/main/java/marketplace_la_u/marketplace_la_u/service/CategoryService.java
@@ -1,5 +1,8 @@
 package marketplace_la_u.marketplace_la_u.service;
-
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import marketplace_la_u.marketplace_la_u.DTO.category.CategoryRequest;
+import marketplace_la_u.marketplace_la_u.DTO.category.CategoryResponse;
 import marketplace_la_u.marketplace_la_u.model.Category;
 import marketplace_la_u.marketplace_la_u.repository.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,47 +11,45 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
+@Transactional
+@RequiredArgsConstructor
 public class CategoryService {
 
-    @Autowired
-    private CategoryRepository repository;
+    private final CategoryRepository repository;
 
-    public List<Category> listar() {
-        return repository.findAll();
+    public List<CategoryResponse> listCategory() {
+        return repository.findAll().stream().map(CategoryResponse::new).toList();
     }
 
-    public Category guardar(Category category) {
-        if (repository.existsByNombre(category.getNombre())) {
+    public CategoryResponse registerCategory(CategoryRequest categoryDto) {
+        if (repository.existsByName(categoryDto.getName())) {
             throw new RuntimeException("La categoría ya existe");
         }
+        Category category = new Category();
+        category.setName(categoryDto.getName());
+        category.setDescripcion(categoryDto.getDescription());
+        Category newCategory = repository.save(category);
 
-        return repository.save(category);
+        return new CategoryResponse(newCategory);
     }
 
-    public Category obtenerPorId(Long id) {
-        return repository.findById(id).orElse(null);
+    public CategoryResponse consultById(Long id) {
+        return repository.findById(id).map(CategoryResponse::new).orElseThrow(() -> new RuntimeException("Categoria no encontrada."));
     }
 
-    public void eliminar(Long id) {
+    public void deleteCategory(Long id){
+        if (!repository.existsById(id)){
+            throw new RuntimeException("Categoria no encontrado");
+        }
         repository.deleteById(id);
     }
 
-    public Category actualizar(Long id, Category category) {
-
-        Category existente = repository.findById(id).orElse(null);
-
-        if (existente == null) {
-            throw new RuntimeException("Categoría no encontrada");
-        }
-
-        if (!existente.getNombre().equals(category.getNombre()) &&
-                repository.existsByNombre(category.getNombre())) {
-            throw new RuntimeException("El nombre ya está en uso");
-        }
-
-        existente.setNombre(category.getNombre());
-        existente.setDescripcion(category.getDescripcion());
-
-        return repository.save(existente);
+    public CategoryResponse updateCategory(Long id, Category nuevosDatos){
+        return repository.findById(id).map( category ->{
+            if(nuevosDatos.getName() != null) category.setName(nuevosDatos.getName());
+            if(nuevosDatos.getDescripcion() != null) category.setDescripcion(nuevosDatos.getDescripcion());
+            return new CategoryResponse(repository.save(category));
+        }).orElseThrow(() -> new RuntimeException("Categoria no encontrada"));
     }
+
 }
